refactor: split api.ts into domain modules under api/
